Final Debug Push to address issues found in #94 with special attention to java classes while views and xml files where left for another branch an PR.

+++ app/src/main/java/com/example/happymeals/InputValidator.java
@@ -59,7 +59,7 @@ public class InputValidator {
         // ensures numbers are entered, numbers are ints, and numbers > 0
         boolean valid = true;
         try {
-            Integer input = Integer.parseInt(field.getText().toString());
+            Double input = Double.parseDouble(field.getText().toString());
             if (input <= 0)
                 valid = false;
         } catch (Exception e) {
